requirements.txt
Babel==2.4.0
blinker==1.3
click==6.7
coverage==4.4.1
decorator==4.0.11
defusedxml==0.5.0
Flask==0.10.1
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==1.0
Flask-WhooshAlchemy==0.8
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pbr==3.0.1
python3-openid==3.1.0
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.10
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.2
Whoosh==2.6.0
WTForms==2.1
